+++ source/gear/_greed.js
@@ -1,13 +1,20 @@
-const { GearTemplate } = require('../classes/index.js');
+const { GearTemplate, GearFamily } = require('../classes/index.js');
 const { addModifier, generateModifierResultLines, combineModifierReceipts } = require('../util/combatantUtil.js');
 
-module.exports = new GearTemplate("Greed",
+//#region Base
+const base = new GearTemplate("Greed",
 	[["use", "Add @{mod0Stacks} @{mod0} and @{mod1Stacks} @{mod1} to all Treasure Elementals with priority"]],
 	"Action",
 	"Unaligned"
-).setEffect((targets, user, adventure) => {
-	const { modifiers: [midas, empowerment] } = module.exports;
+).setEffect(execute, { type: "all", team: "foe" })
+	.setModifiers({ name: "Curse of Midas", stacks: 1 }, { name: "Empowerment", stacks: 20 });
+
+/** @type {typeof base.effect} */
+function execute(targets, user, adventure) {
+	const [midas, empowerment] = base.modifiers;
 	const affectedTargets = targets.filter(target => target.archetype === "Treasure Elemental");
 	return generateModifierResultLines(combineModifierReceipts(addModifier(affectedTargets, empowerment).concat(addModifier(affectedTargets, midas))));
-}, { type: "all", team: "foe" })
-	.setModifiers({ name: "Curse of Midas", stacks: 1 }, { name: "Empowerment", stacks: 20 });
+}
+//#endRegion Base
+
+module.exports = new GearFamily(base, [], true);
